refactor: consolidate web components into layout/ and status/ modules
