+++ apps/web/src/components/layout/event-feed.tsx
